Improved readability, seperated widgets
