+++ lib/main.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
+import 'package:profile_app/profile_field.dart';
+import 'package:profile_app/profile_field_page.dart';
 
 // Creating a User Object
 class User extends ChangeNotifier {
@@ -8,6 +10,7 @@ class User extends ChangeNotifier {
   String phone;
   String email;
   String bio;
+  String image;
 
   User({
     required this.firstName,
@@ -15,35 +18,16 @@ class User extends ChangeNotifier {
     required this.phone,
     required this.email,
     required this.bio,
+    required this.image,
   });
-
-  void setFirstName(String fname) {
-    firstName = fname;
-  }
-
-  void setLastName(String lname) {
-    lastName = lname;
-  }
-
-  void setPhone(String phone) {
-    this.phone = phone;
-  }
-
-  void setEmail(String email) {
-    this.email = email;
-  }
-
-  void setBio(String bio) {
-    this.bio = bio;
-  }
 }
 
 void main() {
-  runApp(const MyApp());
+  runApp(MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({super.key});
+  MyApp({super.key});
 
   final MaterialColor primaryColor =
       const MaterialColor(0xff4169e1, <int, Color>{
@@ -61,15 +45,16 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    // Building Global App Themes
     return MaterialApp(
       title: "Profile App",
       theme: ThemeData(
         useMaterial3: true,
+        scaffoldBackgroundColor: Colors.grey[50],
         colorScheme: ColorScheme.fromSwatch(
             primarySwatch: primaryColor,
-            backgroundColor: const Color.fromRGBO(248, 248, 248, 1),
+            backgroundColor: Colors.white,
             accentColor: Colors.grey[350]),
-        scaffoldBackgroundColor: primaryColor.shade50,
         textTheme: TextTheme(
           displayLarge: GoogleFonts.notoSansKr(
             fontSize: 24,
@@ -103,6 +88,7 @@ class MyApp extends StatelessWidget {
   }
 }
 
+// Application Home Screen
 class HomeScreen extends StatefulWidget {
   const HomeScreen({super.key});
 
@@ -116,17 +102,20 @@ class _HomeScreenState extends State<HomeScreen> {
   @override
   void initState() {
     super.initState();
+
+    // Initialize a User with values in State
     user = User(
-      firstName: 'Micah',
-      lastName: 'Smith',
-      phone: '208-206-5039',
-      email: 'micahsmith@gmail.com',
-      bio:
-          'Hi my name is Mica Smith. I am from Mesa but go to school in Salt Lake City. I make this drive all the time and have plenty',
-    );
+        firstName: 'Micah',
+        lastName: 'Smith',
+        phone: '208-206-5039',
+        email: 'micahsmith@gmail.com',
+        bio:
+            'Hi my name is Mica Smith. I am from Mesa but go to school in Salt Lake City. I make this drive all the time and have plenty',
+        image:
+            'https://firebasestorage.googleapis.com/v0/b/profile-demo-53d30.appspot.com/o/headshot_ph.jpg?alt=media&token=3c1658a9-8378-471f-87a4-9f0d8f4fb8fe');
   }
 
-// Method to update user object in state based on field name
+  // Callback function to update user object in state based on field name
   void updateUser(String updateVal1, String updateVal2, String field) {
     setState(() {
       if (field == "name") {
@@ -142,14 +131,20 @@ class _HomeScreenState extends State<HomeScreen> {
       if (field == "bio") {
         user.bio = updateVal1;
       }
+      if (field == "image") {
+        user.image = updateVal1;
+      }
     });
   }
 
+  // Home Screen Build
   @override
   Widget build(BuildContext context) {
     return Scaffold(
         body: Column(children: <Widget>[
+      // Top Margin
       const SizedBox(height: 60),
+      // Page Header Text
       Text(
         "Edit Profile",
         style: Theme.of(
@@ -157,26 +152,41 @@ class _HomeScreenState extends State<HomeScreen> {
         ).textTheme.displayMedium,
       ),
       const SizedBox(height: 20),
+      // Home Screen Content (Avatar and ListTiles)
       Stack(
         children: [
+          // Circular Profile Image with Border
           CircleAvatar(
             radius: 60,
             backgroundColor: Theme.of(context).colorScheme.primary,
-            child: Padding(
-              padding: const EdgeInsets.all(4),
+            child: const Padding(
+              padding: EdgeInsets.all(4),
               child: ClipOval(
                   child: CircleAvatar(
                 radius: 55,
-                backgroundColor: Colors.deepOrange[100],
+                // backgroundImage: NetworkImage(user.image),
+                backgroundColor: Colors.deepOrange,
               )),
             ),
           ),
+          // Edit Icon Button
           Positioned(
             top: 1,
             right: 1,
             child: ElevatedButton(
+                // When Edit button is pressed, navigate to the 'image' ProfilePagField passing user.image
                 onPressed: () {
-                  debugPrint("Edit Button was clicked");
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (context) => ProfileFieldPage(
+                        initialValue1: user.image,
+                        initialValue2: '',
+                        field: "image",
+                        updateUserCallback: updateUser,
+                      ),
+                    ),
+                  );
                 },
                 style: ElevatedButton.styleFrom(
                   shape: const CircleBorder(),
@@ -188,6 +198,7 @@ class _HomeScreenState extends State<HomeScreen> {
           ),
         ],
       ),
+      // List View with Name, Phone, Email, and Bio Tiles
       Expanded(
         child: Padding(
           padding: const EdgeInsets.all(24.0),
@@ -197,6 +208,7 @@ class _HomeScreenState extends State<HomeScreen> {
                 color: Theme.of(context).colorScheme.secondary,
                 context: context,
                 tiles: [
+                  // Name ListTile navigates to 'name' ProfileFieldPage passing user first and last name
                   ListTile(
                     iconColor: Theme.of(context).colorScheme.secondary,
                     onTap: () {
@@ -212,12 +224,15 @@ class _HomeScreenState extends State<HomeScreen> {
                         ),
                       );
                     },
+                    // Forward arrow Icon
                     trailing: const Icon(Icons.arrow_forward_ios_rounded),
                     contentPadding: EdgeInsets.zero,
+                    // Name Title and Label
                     title: ProfileField(
                         fieldLabel: "Name",
                         fieldValue: "${user.firstName} ${user.lastName}"),
                   ),
+                  // Phone ListTile navigates to 'phone' ProfileFieldPage passing user phone
                   ListTile(
                     iconColor: Theme.of(context).colorScheme.secondary,
                     onTap: () {
@@ -233,11 +248,14 @@ class _HomeScreenState extends State<HomeScreen> {
                         ),
                       );
                     },
+                    // Forward Arrow Icon
                     trailing: const Icon(Icons.arrow_forward_ios_rounded),
                     contentPadding: EdgeInsets.zero,
+                    // Phone Title and Label
                     title: ProfileField(
                         fieldLabel: "Phone", fieldValue: user.phone),
                   ),
+                  // Email ListTile navigates to 'email' ProfileFieldPage passing user email
                   ListTile(
                     iconColor: Theme.of(context).colorScheme.secondary,
                     onTap: () {
@@ -253,11 +271,14 @@ class _HomeScreenState extends State<HomeScreen> {
                         ),
                       );
                     },
+                    // Forward Arrow Icon
                     trailing: const Icon(Icons.arrow_forward_ios_rounded),
                     contentPadding: EdgeInsets.zero,
+                    // Email Title and Label
                     title: ProfileField(
                         fieldLabel: "Email", fieldValue: user.email),
                   ),
+                  // Bio ListTile navigates to 'bio' ProfileFieldPage passing user bio
                   ListTile(
                     iconColor: Theme.of(context).colorScheme.secondary,
                     onTap: () {
@@ -273,8 +294,10 @@ class _HomeScreenState extends State<HomeScreen> {
                         ),
                       );
                     },
+                    // Forward Arrow Icon
                     trailing: const Icon(Icons.arrow_forward_ios_rounded),
                     contentPadding: EdgeInsets.zero,
+                    // Profile Title and Label
                     title: ProfileField(
                         fieldLabel: "Tell us about yourself",
                         fieldValue: user.bio),
@@ -287,260 +310,39 @@ class _HomeScreenState extends State<HomeScreen> {
   }
 }
 
-class ProfileField extends StatelessWidget {
-  const ProfileField(
-      {super.key, required this.fieldLabel, required this.fieldValue});
-
-  final String fieldLabel;
-  final String fieldValue;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.start,
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(
-          fieldLabel,
-          style: Theme.of(
-            context,
-          ).textTheme.labelMedium,
-        ),
-        Text(fieldValue,
-            style: Theme.of(
-              context,
-            ).textTheme.bodyMedium),
-      ],
-    );
-  }
-}
-
-typedef UpdateUserCallback = void Function(
-    String val1, String val2, String field);
-
-class ProfileFieldPage extends StatefulWidget {
-  final String initialValue1, initialValue2, field;
-  final UpdateUserCallback updateUserCallback;
-
-  const ProfileFieldPage(
-      {super.key,
-      required this.initialValue1,
-      required this.initialValue2,
-      required this.field,
-      required this.updateUserCallback});
-
-  @override
-  State<ProfileFieldPage> createState() => _ProfileFieldPageState();
-}
-
-class _ProfileFieldPageState extends State<ProfileFieldPage> {
-  late TextEditingController _controller1;
-  late TextEditingController _controller2;
-  late GlobalKey<FormState> _formKey;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller1 = TextEditingController(text: widget.initialValue1);
-    _controller2 = TextEditingController(text: widget.initialValue2);
-    _formKey = GlobalKey<FormState>();
-  }
+// Loading Widget
+class Loading extends StatelessWidget {
+  const Loading({super.key});
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(),
-      body: Padding(
-          padding: const EdgeInsets.all(32),
-          child: Form(
-              key: _formKey,
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.start,
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  if (widget.field == "name")
-                    Text(
-                      "What's your name?",
-                      style: Theme.of(context).textTheme.displayLarge,
-                    ),
-                  if (widget.field == "phone")
-                    Text(
-                      "What's your phone number?",
-                      style: Theme.of(context).textTheme.displayLarge,
-                    ),
-                  if (widget.field == "email")
-                    Text(
-                      "What's your email?",
-                      style: Theme.of(context).textTheme.displayLarge,
-                    ),
-                  if (widget.field == "bio")
-                    Text(
-                      "What type of passenger are you?",
-                      style: Theme.of(context).textTheme.displayLarge,
-                    ),
-                  const SizedBox(
-                    height: 30,
-                  ),
-                  Row(
-                    children: <Widget>[
-                      if (widget.field == "name")
-                        Expanded(
-                          child: Container(
-                            decoration: BoxDecoration(
-                                border: Border.all(
-                                    color: Theme.of(context)
-                                        .colorScheme
-                                        .secondary)),
-                            child: TextFormField(
-                              keyboardType: TextInputType.name,
-                              controller: _controller1,
-                              decoration: InputDecoration(
-                                border: InputBorder.none,
-                                labelText: "First Name",
-                                labelStyle:
-                                    Theme.of(context).textTheme.labelSmall,
-                                contentPadding: const EdgeInsets.all(8),
-                              ),
-                              style: Theme.of(context).textTheme.bodySmall,
-                              validator: (value) {
-                                if (value == null || value.isEmpty) {
-                                  return 'Field cannot be empty';
-                                }
-                                return null;
-                              },
-                            ),
-                          ),
-                        ),
-                      if (widget.field == "phone")
-                        Expanded(
-                          child: Container(
-                            decoration: BoxDecoration(
-                                border: Border.all(
-                                    color: Theme.of(context)
-                                        .colorScheme
-                                        .secondary)),
-                            child: TextFormField(
-                              keyboardType: TextInputType.phone,
-                              controller: _controller1,
-                              decoration: InputDecoration(
-                                border: InputBorder.none,
-                                labelText: "Your phone number",
-                                labelStyle:
-                                    Theme.of(context).textTheme.labelSmall,
-                                contentPadding: const EdgeInsets.all(8),
-                              ),
-                              style: Theme.of(context).textTheme.bodySmall,
-                              validator: (value) {
-                                if (value == null || value.isEmpty) {
-                                  return 'Field cannot be empty';
-                                }
-                                return null;
-                              },
-                            ),
-                          ),
-                        ),
-                      if (widget.field == "email")
-                        Expanded(
-                          child: Container(
-                            decoration: BoxDecoration(
-                                border: Border.all(
-                                    color: Theme.of(context)
-                                        .colorScheme
-                                        .secondary)),
-                            child: TextFormField(
-                              keyboardType: TextInputType.emailAddress,
-                              controller: _controller1,
-                              decoration: InputDecoration(
-                                border: InputBorder.none,
-                                labelText: "Your email address",
-                                labelStyle:
-                                    Theme.of(context).textTheme.labelSmall,
-                                contentPadding: const EdgeInsets.all(8),
-                              ),
-                              style: Theme.of(context).textTheme.bodySmall,
-                              validator: (value) {
-                                if (value == null || value.isEmpty) {
-                                  return 'Field cannot be empty';
-                                }
-                                return null;
-                              },
-                            ),
-                          ),
-                        ),
-                      if (widget.field == "bio")
-                        Expanded(
-                          child: Container(
-                            decoration: BoxDecoration(
-                                border: Border.all(
-                                    color: Theme.of(context)
-                                        .colorScheme
-                                        .secondary)),
-                            child: TextFormField(
-                              keyboardType: TextInputType.multiline,
-                              maxLines: null,
-                              controller: _controller1,
-                              decoration: InputDecoration(
-                                border: InputBorder.none,
-                                hintText: "Write a little bit about yourself",
-                                labelStyle:
-                                    Theme.of(context).textTheme.labelSmall,
-                                contentPadding: const EdgeInsets.all(8),
-                              ),
-                              style: Theme.of(context).textTheme.bodySmall,
-                              validator: (value) {
-                                if (value == null || value.isEmpty) {
-                                  return 'Field cannot be empty';
-                                }
-                                return null;
-                              },
-                            ),
-                          ),
-                        ),
-                      const SizedBox(width: 20),
-                      if (widget.field == "name")
-                        Flexible(
-                          child: Container(
-                            decoration: BoxDecoration(
-                                border: Border.all(
-                                    color: Theme.of(context)
-                                        .colorScheme
-                                        .secondary)),
-                            child: TextFormField(
-                              keyboardType: TextInputType.name,
-                              controller: _controller2,
-                              decoration: InputDecoration(
-                                border: InputBorder.none,
-                                labelText: "Last Name",
-                                labelStyle:
-                                    Theme.of(context).textTheme.labelSmall,
-                                contentPadding: const EdgeInsets.all(8),
-                              ),
-                              style: Theme.of(context).textTheme.bodySmall,
-                              validator: (value) {
-                                if (value == null || value.isEmpty) {
-                                  return 'Field cannot be empty';
-                                }
-                                return null;
-                              },
-                            ),
-                          ),
-                        ),
-                    ],
-                  ),
-                  const SizedBox(height: 16.0),
-                  ElevatedButton(
-                    onPressed: () {
-                      if (_formKey.currentState?.validate() ?? false) {
-                        widget.updateUserCallback(
-                            _controller1.text, _controller2.text, widget.field);
-                        // Validate successful, pass the updated value back
-                        Navigator.pop(context);
-                      }
-                    },
-                    child: const Text('Update'),
-                  ),
-                ],
-              ))),
+      appBar: AppBar(title: const Text("Loading...")),
+      body: const Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [Text("Loading your profile")],
+      ),
+    );
+  }
+}
+
+// Application Error Screen
+class ErrorScreen extends StatelessWidget {
+  const ErrorScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text("Error Screen")),
+      body: const Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Text(
+            "There has been an error",
+            style: TextStyle(color: Colors.red),
+          )
+        ],
+      ),
     );
   }
 }
