+++ pubspec.yaml
@@ -24,7 +24,6 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.2
   provider: ^6.1.1
   google_fonts: ^6.1.0
